Create package for a preconfigured logger
I'll use this as a sort of integration test. It also serves as a quick
get-started package for those who don't use the library yet, or my
future self that is too lazy to configure the logger for a new project.

It will support browsers and node environments. It should be
a reasonable replacement for `debug`.