Rename terminal backend to specify ANSI
The terminal backend is very specialized to handle ANSI codes. I tried
making an option that disabled colors and quickly realized it was better
to create a plaintext new backend. Some content that relies on color
ques requires different representation if you're not in a TTY.

This also makes it explicit that older versions of Windows might not
support the output, although I think anything pre-10 is about to be
EOL'd.