Only use ANSI control codes in TTYs
This disables ANSI control codes if your terminal doesn't support 3-bit
colors or if the log destination isn't a TTY. It falls back to using the
stream backend over `process.stderr`.

This is standard behavior. Run a program directly and you'll see colors.
Pipe it to `cat` and it switches to machine-optimized output.