Stage version changes
I've been playing with the `yarn version` plugin a bit more. I think
I understand how it can fit my workflow. Recursively bumping versions
requires fixed version numbers, so I've changed that to match and staged
all the modified packages for a new minor bump.

It would be nice to get this automated. It's enough of a pain that
I don't think I'll remember the process coming back to it 6 months from
now.