Tree shake unused ANSI codes from bundles
Rollup is smart enough to detect that I'm only using half the colors
I defined. I'd like to keep the others, but if I tweak the way I export,
I don't have to pay the distribution expense.

TBH I'm just pinching pennies here. It was fun to get it working.