Migrate all plugins to functions
The `LogProcessor` type used to be any object with a `processLog(...)`
method. Now, it is any function that implements `Log => void`.

The main reason for this is bundle size: if you look in the bundled
assets, you'll see class symbols unmangled (e.g. `processLog`,
`loadPattern`, `this.stream`). It adds up. I want to keep Holz as light
as possible in the browser.

The second reason is composition: if plugins are just functions, it
becomes easier to leverage functional composition. There are plenty of
conventional tools that operate on functions, while there are almost
none that operate on Holz plugins (e.g. `_.pipe` doesn't work on
a `StreamBackend`).