Generate TS declaration files with each package
Previously I was pointing TypeScript at the source code for types. It
worked well enough as long as your compiler was closely aligned with the
version. Holz is stable enough that I don't publish often, so this
wasn't a good strategy.

The tooling has gotten better and maintaining build steps for
declaration files has gotten much easier. This should increase the
stability of the type defs and also help with TS performance.

Note: I had to export the `Logger` type in order to generate the
`@holz/logger` bundle. Otherwise the generator would've tried to inline
the logger type along with its private methods which is unsupported.

This seems like a good change regardless.