Add `logger.withMiddleware` method
I've wanted this for a while but hadn't gotten around to it. It supports
pre-processing logs before sending them to the default plugins. I find
it's useful for adding default context to a logger (such as in a class
with important IDs) or enabling a verbose style of logs just for that
module (such as only printing traces if an attribute is set).