Replace trace color with bright white
There's a slight risk it might not be supported in older terminals. In
practice that might just be older CI systems. It will fall back to the
current text color, which is likely close to white anyway.

While revisiting terminal color support I noticed my color check was
a bit wacky. I mistook that for color depth instead of number of
available colors. Now it's using the appropriate API.